Only support 10.9 (Mavericks) and up. Currently checked in: booter, kernel & version plist.
Drop Clang support.
Proper QPI initializing to inject/patch smbios type132 as needed. Thx
to @theracermaster!

Added more disk template vars, ex:

<key>GUI</key>
<dict>
  <key>DarwinRecoveryDiskTemplate</key>
  <string>$platform -$label- ($major.$minor.$revision | $build) on
'$path'</string>
  ...

Will resulting: Darwin -Recovery- (10.12.0 | 16A323) on Recovery HD

Available DiskTemplate keys:
DarwinDiskTemplate, DarwinRecoveryDiskTemplate,
DarwinInstallerDiskTemplate, LinuxDiskTemplate, WindowsDiskTemplate

Available vars/placeholder:
$label, $path, $platform, $version, $build, $major, $minor, $revision

+++ Library/Platform/Events.c
@@ -29,8 +29,8 @@ SaveDarwinLog () {
   UINT32    PropSize;
 
   if (DTLookupEntry (NULL, "/efi/platform", &PlatformEntry) == kSuccess) {
-    if (DTGetProperty(PlatformEntry, DATAHUB_LOG, &PropValue, &PropSize) == kSuccess) {
-      CONSTRAIN_MAX (PropSize, GetMemLogLen ());
+    if (DTGetProperty (PlatformEntry, DATAHUB_LOG, &PropValue, &PropSize) == kSuccess) {
+      CONSTRAIN_MAX (PropSize, (UINT32)GetMemLogLen ());
       CopyMem (PropValue, GetMemLogBuffer (), PropSize);
     }
   }
